i9300: libsec-ril: Disable set_wakelock calls
Hexedited the blob and replaced all calls to set_wakelock
with NOP. The wakelock is done in RIL_Service.

Since Android Q android.system.suspend@1.0::IWakeLock has been
binderized too. The libsec-ril's internal method set_wakelock
is internally called and calls via binder to set the wakelock.
But the initial call is done via binder in the first place and
is being handled at this point. The set_wakelock call
is succesfully handled but overwrites internally mOut in
IPCThreadState (libhwbinder). The initial call fails because its
message is of set_wakelock. DataSize of the mOut-Parcel mismatches
and bails out and causes the crash of the RIL_service.

Change-Id: I03797ad74a19ce3061b58a474da9faa735bf91c6